# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Target cloud provider
 */
export type CloudProvider = typeof CloudProvider[keyof typeof CloudProvider];


export const CloudProvider = {
  aws: 'aws',
  azure: 'azure',
  gcp: 'gcp',
} as const;

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  /** The user's personal workspace */
  workspaceId: number;
}

export interface RegisterInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

/**
 * A regenerable content section of an architecture
 */
export type ArchitectureSection = typeof ArchitectureSection[keyof typeof ArchitectureSection];


export const ArchitectureSection = {
  diagram: 'diagram',
  terraform: 'terraform',
  costEstimate: 'costEstimate',
  securityRecommendations: 'securityRecommendations',
  highAvailabilityPlan: 'highAvailabilityPlan',
  databaseRecommendation: 'databaseRecommendation',
  kubernetesDeployment: 'kubernetesDeployment',
  cicdPipeline: 'cicdPipeline',
  monitoringSetup: 'monitoringSetup',
  disasterRecovery: 'disasterRecovery',
  threatModel: 'threatModel',
} as const;

export interface RegenerateSectionInput {
  section: ArchitectureSection;
  /** Optional extra guidance for the regeneration */
  instructions?: string;
}

export interface ArchitectureVersionSummary {
  id: number;
  architectureId: number;
  versionNumber: number;
  reason: string;
  changedFields: string[];
  /** @nullable */
  editorName: string | null;
  createdAt: string;
}

/**
 * Content fields as they were before the change
 */
export type ArchitectureVersionSnapshot = {[key: string]: string};

export type ArchitectureVersion = ArchitectureVersionSummary & {
  /** Content fields as they were before the change */
  snapshot: ArchitectureVersionSnapshot;
};

export type AuditLogEntryMetadata = { [key: string]: unknown };

export interface AuditLogEntry {
  id: number;
  actionType: string;
  entityType: string;
  /** @nullable */
  entityId: number | null;
  metadata: AuditLogEntryMetadata;
  /** @nullable */
  userName: string | null;
  createdAt: string;
}

export interface Architecture {
  id: number;
  title: string;
  requirements: string;
  provider: CloudProvider;
  /** Mermaid diagram source */
  diagram: string;
  /** Terraform HCL content */
  terraform: string;
  costEstimate: string;
  securityRecommendations: string;
  highAvailabilityPlan: string;
  databaseRecommendation: string;
  kubernetesDeployment: string;
  cicdPipeline: string;
  monitoringSetup: string;
  disasterRecovery: string;
  /** STRIDE threat model as a JSON document */
  threatModel: string;
  createdAt: string;
  updatedAt: string;
}

export interface ArchitectureInput {
  title: string;
  requirements: string;
  provider?: CloudProvider;
  diagram: string;
  terraform: string;
  costEstimate: string;
  securityRecommendations: string;
  highAvailabilityPlan: string;
  databaseRecommendation: string;
  kubernetesDeployment: string;
  cicdPipeline: string;
  monitoringSetup: string;
  disasterRecovery: string;
  /** STRIDE threat model as a JSON document */
  threatModel?: string;
}

export interface ArchitectureUpdate {
  title?: string;
  requirements?: string;
  provider?: CloudProvider;
  diagram?: string;
  terraform?: string;
  costEstimate?: string;
  securityRecommendations?: string;
  highAvailabilityPlan?: string;
  databaseRecommendation?: string;
  kubernetesDeployment?: string;
  cicdPipeline?: string;
  monitoringSetup?: string;
  disasterRecovery?: string;
  threatModel?: string;
}

export interface GenerateArchitectureInput {
  /** Natural language description of the application requirements */
  requirements: string;
  provider?: CloudProvider;
}

/**
 * Infrastructure code format
 */
export type ValidateInfrastructureInputFormat = typeof ValidateInfrastructureInputFormat[keyof typeof ValidateInfrastructureInputFormat];


export const ValidateInfrastructureInputFormat = {
  terraform: 'terraform',
  cloudformation: 'cloudformation',
  kubernetes: 'kubernetes',
} as const;

export interface ValidateInfrastructureInput {
  /** Raw infrastructure code to analyze (uploaded or pasted) */
  source: string;
  /** Infrastructure code format */
  format: ValidateInfrastructureInputFormat;
}

export interface ArchitectureStats {
  totalCount: number;
  /** Architectures created in the last 7 days */
  recentCount: number;
  /** @nullable */
  latestTitle: string | null;
}

export interface ApiError {
  error: string;
}

export type ListAuditLogsParams = {
/**
 * Limit results to one architecture
 */
architectureId?: number;
limit?: number;
};