# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created; session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in; session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /architectures:
    get:
      operationId: listArchitectures
      tags: [architectures]
      summary: List all saved architectures
      responses:
        "200":
          description: List of architectures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Architecture"
    post:
      operationId: saveArchitecture
      tags: [architectures]
      summary: Save a generated architecture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchitectureInput"
      responses:
        "201":
          description: Saved architecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architecture"
  /architectures/{id}:
    get:
      operationId: getArchitecture
      tags: [architectures]
      summary: Get a single architecture by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Architecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateArchitecture
      tags: [architectures]
      summary: Update an architecture (e.g. rename)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchitectureUpdate"
      responses:
        "200":
          description: Updated architecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteArchitecture
      tags: [architectures]
      summary: Delete an architecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /architectures/generate:
    post:
      operationId: generateArchitecture
      tags: [architectures]
      summary: Generate an AWS architecture from natural language requirements (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateArchitectureInput"
      responses:
        "200":
          description: SSE stream of architecture sections
          content:
            text/event-stream: {}
  /architectures/{id}/terraform:
    get:
      operationId: downloadTerraform
      tags: [architectures]
      summary: Download Terraform files for an architecture as plain text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Terraform HCL content
          content:
            text/plain: {}
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /architectures/{id}/regenerate-section:
    post:
      operationId: regenerateSection
      tags: [architectures]
      summary: Regenerate one section of a saved architecture (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateSectionInput"
      responses:
        "200":
          description: SSE stream ending with the regenerated section content
          content:
            text/event-stream: {}
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    CloudProvider:
      type: string
      description: Target cloud provider
      enum: [aws, azure, gcp]
    AuthUser:
      type: object
      required: [id, email, name, workspaceId]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        workspaceId:
          type: integer
          description: The user's personal workspace
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ArchitectureSection:
      type: string
      description: A regenerable content section of an architecture
      enum:
        - diagram
        - terraform
        - costEstimate
        - securityRecommendations
        - highAvailabilityPlan
        - databaseRecommendation
        - kubernetesDeployment
        - cicdPipeline
        - monitoringSetup
        - disasterRecovery
        - threatModel
    RegenerateSectionInput:
      type: object
      required: [section]
      properties:
        section:
          $ref: "#/components/schemas/ArchitectureSection"
        instructions:
          type: string
          description: Optional extra guidance for the regeneration
...
    Architecture:
      type: object
      required: [id, title, requirements, provider, diagram, terraform, costEstimate, securityRecommendations, highAvailabilityPlan, databaseRecommendation, kubernetesDeployment, cicdPipeline, monitoringSetup, disasterRecovery, threatModel, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        requirements:
          type: string
        provider:
          $ref: "#/components/schemas/CloudProvider"
        diagram:
          type: string
          description: Mermaid diagram source
        terraform:
          type: string
          description: Terraform HCL content
        costEstimate:
          type: string
        securityRecommendations:
          type: string
        highAvailabilityPlan:
          type: string
        databaseRecommendation:
          type: string
        kubernetesDeployment:
          type: string
        cicdPipeline:
          type: string
        monitoringSetup:
          type: string
        disasterRecovery:
          type: string
        threatModel:
          type: string
          description: STRIDE threat model as a JSON document
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArchitectureInput:
      type: object
      required: [title, requirements, diagram, terraform, costEstimate, securityRecommendations, highAvailabilityPlan, databaseRecommendation, kubernetesDeployment, cicdPipeline, monitoringSetup, disasterRecovery]
      properties:
        title:
          type: string
        requirements:
          type: string
        provider:
          $ref: "#/components/schemas/CloudProvider"
        diagram:
          type: string
        terraform:
          type: string
        costEstimate:
          type: string
        securityRecommendations:
          type: string
        highAvailabilityPlan:
          type: string
        databaseRecommendation:
          type: string
        kubernetesDeployment:
          type: string
        cicdPipeline:
          type: string
        monitoringSetup:
          type: string
        disasterRecovery:
          type: string
        threatModel:
          type: string
          description: STRIDE threat model as a JSON document
    ArchitectureUpdate:
      type: object
      properties:
        title:
          type: string
        requirements:
          type: string
        provider:
          $ref: "#/components/schemas/CloudProvider"
        diagram:
          type: string
        terraform:
          type: string
        costEstimate:
          type: string
        securityRecommendations:
          type: string
        highAvailabilityPlan:
          type: string
        databaseRecommendation:
          type: string
        kubernetesDeployment:
          type: string
        cicdPipeline:
          type: string
        monitoringSetup:
          type: string
        disasterRecovery:
          type: string
        threatModel:
          type: string
    GenerateArchitectureInput:
      type: object
      required: [requirements]
      properties:
        requirements:
          type: string
          description: Natural language description of the application requirements
        provider:
          $ref: "#/components/schemas/CloudProvider"
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string