# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  /** The user's personal workspace */
  workspaceId: number;
}